Dados dos principais reservatórios brasileiros do módulo Nordeste e Semiárido do SAR

Dados dos principais reservatórios brasileiros do módulo Nordeste e Semiárido do SAR

reservatorio_nordeste_semiarido(
  codigo_reservatorio,
  data_inicial = "1980-01-01",
  data_final = Sys.Date()
)

Arguments

codigo_reservatorio
Codigo do reservatorio a ser buscado, (voce pode consultar a tabela com os codigos chamando tabela_reservatorios).

data_inicial
Data inicial do registro. Se for anterior a data de funcionamento, retorna a data de registro mais antiga.

data_final
Data final do registro. Pode ser usado Sys.Date() para retornar a data de registro mais recente.

Examples

reservatorio_nordeste_semiarido(12354, "2012-09-10", Sys.Date())
#> ℹ Buscando as informações do reservatório PONTO NOVO
#> Eita, isso não vai demorar (tanto) ...
#> Você já regou as plantas hoje? 
#> ✔ O que achou?
#> # A tibble: 1,246 × 7
#>    data       estado reservatorio capacidade_hm3 cota_m volume_hm3
#>    <date>     <fct>  <fct>                 <dbl>  <dbl>      <dbl>
#>  1 2012-09-10 Bahia  PONTO NOVO             38.9   355.      10.6 
#>  2 2012-11-26 Bahia  PONTO NOVO             38.9   353.       6.25
#>  3 2012-12-10 Bahia  PONTO NOVO             38.9   353.       5.18
#>  4 2012-12-27 Bahia  PONTO NOVO             38.9   353.       5.58
#>  5 2013-01-07 Bahia  PONTO NOVO             38.9   353.       5.18
#>  6 2013-01-14 Bahia  PONTO NOVO             38.9   353.       4.89
#>  7 2013-02-04 Bahia  PONTO NOVO             38.9   356.      14.9 
#>  8 2013-02-18 Bahia  PONTO NOVO             38.9   356.      13.7 
#>  9 2013-02-25 Bahia  PONTO NOVO             38.9   355.      12.8 
#> 10 2013-03-18 Bahia  PONTO NOVO             38.9   355.       9.99
#> # … with 1,236 more rows, and 1 more variable: volume_percentual <dbl>