Tabela com dados dos reservatorios brasileiros

Um dataset contendo os atributos dos principais reservatorios brasileiros.

tabela_reservatorios()

Format

Um dataframe com 708 observacoes e 5 variaveis

sistema

sistema o qual o reservatorio pertence

codigo

o codigo do reservatorio segundo o SAR

estado_1

estado o qual o reservatorio pertence

estado_2

se o reservatorio fica entre estados, o segundo estado

reservatorio

o nome do reservatorio

Source

https://www.ana.gov.br/sar0/. obtido em 2021-08-22

Examples

tabela_reservatorios()

#> # A tibble: 710 x 8
#>    sistema            codigo reservatorio    bacia rio   estado_1 estado_2   ano
#>    <fct>               <int> <fct>           <fct> <lgl> <fct>    <fct>    <int>
#>  1 nordeste_semiarido  12001 25 DE MARÇO     NA    NA    RN       NA          NA
#>  2 nordeste_semiarido  12002 ABÓBORAS        NA    NA    PE       NA          NA
#>  3 nordeste_semiarido  12003 ACARAPE DO MEIO NA    NA    CE       NA          NA
#>  4 nordeste_semiarido  12004 ACARAÚ MIRIM    NA    NA    CE       NA        1907
#>  5 nordeste_semiarido  12005 ACAUÃ           NA    NA    PB       NA          NA
#>  6 nordeste_semiarido  12006 ADAUTO BEZERRA  NA    NA    CE       NA          NA
#>  7 nordeste_semiarido  12007 ADUSTINA        NA    NA    BA       NA          NA
#>  8 nordeste_semiarido  12008 ÁGUA FRIA II    NA    NA    BA       NA          NA
#>  9 nordeste_semiarido  12009 AIPIM           NA    NA    BA       NA          NA
#> 10 nordeste_semiarido  12010 ALAGADIÇO       NA    NA    SE       NA          NA
#> # ... with 700 more rows