Tabela com dados dos reservatorios brasileiros apresentados pela ONS

Um dataset contendo os atributos dos principais reservatorios brasileiros apresentados pela ONS.

tabela_reservatorios_ONS()

Format

Um dataframe com 39 observacoes e 6 variaveis

reservatorio

o nome do reservatorio

codigo

o codigo do reservatorio segundo o SAR

bacia

bacia hidrológica a qual o reservatorio pertence

subsistema

subsistema-ONS a qual o reservatorio pertence

estado_1

estado o qual o reservatorio pertence

estado_2

se o reservatorio fica entre estados, o segundo estado

Source

http://www.ons.org.br/paginas/energia-agora/reservatorios. obtido em 2021-08-30

Examples

tabela_reservatorios_ONS()

#> # A tibble: 39 x 6
#>    subsistema             reservatorio           codigo bacia  estado_1 estado_2
#>    <fct>                  <fct>                   <int> <fct>  <fct>    <fct>   
#>  1 Nordeste               IRAPE                   19115 JEQUI~ MG       NA      
#>  2 Nordeste               SOBRADINHO              19121 SAO F~ BA       NA      
#>  3 Nordeste               TRES MARIAS             19119 SAO F~ MG       NA      
#>  4 Nordeste               LUIZ GONZAGA/ITAPARICA  19122 SAO F~ BA       PE      
#>  5 Norte                  BALBINA                 19141 AMAZO~ AM       NA      
#>  6 Norte                  SERRA DA MESA           19128 TOCAN~ GO       NA      
#>  7 Norte                  TUCURUI                 19134 TOCAN~ PA       NA      
#>  8 Sudeste / Centro-Oeste AGUA VERMELHA           19015 GRANDE MG       SP      
#>  9 Sudeste / Centro-Oeste FURNAS                  19004 GRANDE MG       NA      
#> 10 Sudeste / Centro-Oeste MASCARENHAS DE MORAES   19005 GRANDE MG       NA      
#> # ... with 29 more rows