Tabela com dados dos reservatorios brasileiros

Um dataset contendo os atributos dos principais reservatorios brasileiros.

tabela_reservatorios()

Format

Um dataframe com 805 observacoes e 12 variaveis

sistema

sistema o qual o reservatorio pertence segundo o SAR

codigo

o codigo do reservatorio segundo o SAR

reservatorio

o nome do reservatorio

res_latitude

latitude do reservatorio

res_longitude

longitude do reservatorio

municipio

municipio do reservatorio

codigo_municipio_ibge

codigo do municipio na base do ibge

estado_nome

estado em forma de nome

estado_sigla

estado em forma de sigla

rio

nome do rio

sub_bacia

sub-bacia do rio

bacia

bacia hidrográfica do rio

Source

ANA (Antônio Augusto, Diego Pena e Théo Albuquerque), editado por Bruno Mioto. obtido em 2021-08-22

Examples

tabela_reservatorios()
#> # A tibble: 804 x 12
#>    sistema            codigo reservatorio  res_latitude res_longitude municipio 
#>    <chr>               <int> <chr>                <dbl>         <dbl> <chr>     
#>  1 nordeste_semiarido  12001 25 DE MARÇO          -6.11         -38.2 PAU DOS F~
#>  2 nordeste_semiarido  12002 ABÓBORAS             -8.09         -39.4 PARNAMIRIM
#>  3 nordeste_semiarido  12003 ACARAPE DO M~        -4.19         -38.8 REDENÇÃO  
#>  4 nordeste_semiarido  12004 ACARAÚ MIRIM         -3.51         -40.3 MASSAPÊ   
#>  5 nordeste_semiarido  12005 ACAUÃ                -7.44         -35.6 ITATUBA   
#>  6 nordeste_semiarido  12006 ADAUTO BEZER~        -6.04         -38.4 PEREIRO   
#>  7 nordeste_semiarido  12007 ADUSTINA            -10.6          -38.1 ADUSTINA  
#>  8 nordeste_semiarido  12008 ÁGUA FRIA II        -14.9          -40.6 BARRA DO ~
#>  9 nordeste_semiarido  12009 AIPIM               -10.6          -40.3 ANTONIO G~
#> 10 nordeste_semiarido  12010 ALAGADIÇO           -10.5          -37.6 FREI PAULO
#> # ... with 794 more rows, and 6 more variables: codigo_municipio_ibge <int>,
#> #   estado_nome <chr>, estado_sigla <chr>, rio <chr>, sub_bacia <chr>,
#> #   bacia <chr>